+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Unreleased
 
-- **[UPDATE]** `Text` component is now replacing \n by `<br/>` tags
+- **[UPDATE]** `Text` component is now replacing \n by `<br/>` tags by default
   [...]
 
 # v0.10.0 (27/08/2018)